Add ESLint (type-checked) and fix every finding; lint runs in CI
Setup: ESLint 10 flat config with typescript-eslint's recommendedTypeChecked
set plus the react-hooks v7 rules; `void promise` stays the sanctioned
fire-and-forget marker (no-floating-promises ignoreVoid). npm run lint,
wired into CI before the build.

The initial run found 36 problems, now zero:
- 13 auto-fixed unnecessary type assertions
- any-leaks at the ONNX/JSON boundaries typed properly (embedder,
  embedderNode, store, tts.worker, verify-tokenizer) and a dead
  assignment removed
- React fixes: HeroChat's reduced-motion transcript moved to lazy state
  init; the mobile menu closes from item event handlers instead of a
  route-change effect; SettingsMenu funnels every close through one
  handler so a pending confirmation can't survive reopening; the
  latest-callbacks ref is written in an effect (sanctioned pattern)
- three intentional patterns carry targeted eslint-disable directives
  with reasons (progress-tick ETA estimator, one-shot history restore,
  deliberately narrowed persona deps)

Typecheck, build, and eval (12/12) green.

+++ scripts/verify-turn.ts
@@ -21,7 +21,7 @@ const sig = (x: number): number => 1 / (1 + Math.exp(-x))
 async function refFeatures(audio: Float32Array): Promise<Float32Array> {
   const out = (await (proc as unknown as (a: Float32Array) => Promise<{ input_features: { data: Float32Array; dims: number[] } }>)(audio)).input_features
   const frames = out.dims[out.dims.length - 1]
-  if (frames === 800) return out.data as Float32Array
+  if (frames === 800) return out.data
   const fixed = new Float32Array(80 * 800) // pinFeatures: right-pad/truncate to 800
   const copy = Math.min(frames, 800)
   for (let m = 0; m < 80; m++) for (let f = 0; f < copy; f++) fixed[m * 800 + f] = out.data[m * frames + f]
